Move apps into launcher package to reflect current dependency structure
